Update patch set 1
Patch Set 1: Code-Review+1

This driver is adapted to the Huashan AS3677, using the adc_als_value from the AS3677 kernel driver, and allows us to have a proper ALS adaptive brightness fix.

Even though the code is similar to the AS3676, it is cleaner to have an adapted driver than can then be edited if needed. It also gives the hardware review the real component identification and not the "AS3676" which the Huashan doesn't have.

Patch-set: 1