Update patch set 2
Patch Set 2:

Just for reference, MONOTONIC and BOOTTIME are giving very similar timestamps after Android boot, which gives a properly working autorotation.

But as soon as the phone goes into deepsleep, the MONOTONIC stops and the timestamps read later are in the past.
BOOTTIME isn't impacted by the deepsleep times.

For example, MONOTONIC 491736689047, BOOTTIME 854708787839.

Patch-set: 2